+++ super_test_engine/lib/super_test_engine/copy_app_templates/controllers/favorite_things_controller.rb
@@ -22,7 +22,7 @@ module Admin
       end
 
       def display_schema(action:)
-        Super::Display.new(action: action) do |fields, type|
+        Super::Display.new do |fields, type|
           fields[:name] = type.string
           fields[:member] = type.manual { |member| "#{member.name} (member ##{member.id})" }
         end