+++ super_test_engine/lib/super_test_engine/copy_app_templates/controllers/members_controller.rb
@@ -27,7 +27,7 @@ module Admin
       end
 
       def display_schema(action:)
-        Super::Display.new(action: action) do |fields, type|
+        Super::Display.new do |fields, type|
           fields[:name] = type.string
           fields[:rank] = type.string
           fields[:position] = type.string