+++ test/lib/super/display/guesser_test.rb
@@ -48,18 +48,20 @@ module Super
     end
 
     def test_show
-      display = Display.new(action: action!(:show)) do |fields, type|
+      display = Display.new do |fields, type|
         Display::Guesser.new(model: Olympian, action: action!(:show), fields: fields, type: type).call
       end
+      display.apply(action: action!(:show))
 
       assert_equal(
         %w[id name zeus poseidon athena aphrodite artemis ares dionysus hermes demeter created_at updated_at],
         display.each_attribute_name.to_a
       )
 
-      display = Display.new(action: action!(:show)) do |fields, type|
+      display = Display.new do |fields, type|
         Display::Guesser.new(model: Chthonic, action: action!(:show), fields: fields, type: type).call
       end
+      display.apply(action: action!(:show))
 
       assert_equal(
         %w[id hades nyx charon],
@@ -68,18 +70,20 @@ module Super
     end
 
     def test_index
-      display = Display.new(action: action!(:index)) do |fields, type|
+      display = Display.new do |fields, type|
         Display::Guesser.new(model: Olympian, action: action!(:index), fields: fields, type: type).call
       end
+      display.apply(action: action!(:index))
 
       assert_equal(
         %w[id name zeus poseidon athena] + [:actions],
         display.each_attribute_name.to_a
       )
 
-      display = Display.new(action: action!(:index)) do |fields, type|
+      display = Display.new do |fields, type|
         Display::Guesser.new(model: Chthonic, action: action!(:index), fields: fields, type: type).call
       end
+      display.apply(action: action!(:index))
 
       assert_equal(
         %w[id hades nyx charon] + [:actions],